backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
supabase==2.9.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
python-dotenv==1.0.1
anthropic==0.39.0
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
